requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.32.4 